requirements.txt
python-telegram-bot==20.7
mcrcon==0.7.1 